# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de manipulação de produtos",
        "description": "Essa API tem como objetivo a manipulação dos produtos",
        "termsOfService": "http://localhost:3000/terms",
        "contact": {
            "email": "de@demostenes.com.br"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000/api-docs",
            "description": "API Test"
        },
        {
            "url": "https://api.example.com/api/v1",
            "description": "API Production"
        }
    ],
    "paths": {
        "/products": {
            "post": {
                "summary": "Cadastro de produto",
                "description": "Essa rota será responsável por cadastrar um novo produto",
                "tags": [
                    "Products"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            },
                            "examples": {
                                "product": {
                                    "value": {
                                        "name": "Mouse USB",
                                        "description": "Mouse USB RGB",
                                        "price": 300
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Product already exists!"
                    },
                    "200": {
                        "descritpion": "OK",
                        "content": {
                            "application/json": {
                                "type": "object",
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                }
            }
        },
        "/products/findByName": {
            "get": {
                "description": "Busca de um produto pelo nome",
                "summary": "Busc de um produto pelo nome",
                "tags": [
                    "Products"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nome do produto para busca",
                        "required": true
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Product already exists!"
                    },
                    "200": {
                        "descritpion": "OK",
                        "content": {
                            "application/json": {
                                "type": "object",
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "description": "Busca de um produto pelo id",
                "summary": "Busc de um produto pelo id",
                "tags": [
                    "Products"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Nome do produto pelo id",
                        "required": true
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Product already exists!"
                    },
                    "200": {
                        "descritpion": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Alterar um produto pelo id",
                "summary": "Alterar um produto pelo id",
                "tags": [
                    "Products"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Nome do produto pelo id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            },
                            "examples": {
                                "product": {
                                    "value": {
                                        "name": "Mouse USB sem fio",
                                        "description": "Mouse USB RGB sem fio",
                                        "price": 129
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Product already exists!"
                    },
                    "200": {
                        "descritpion": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Product": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Notebook"
                    },
                    "description": {
                        "type": "string",
                        "example": "Notebook de escritório"
                    },
                    "price": {
                        "type": "number",
                        "example": 1000
                    },
                    "id": {
                        "type": "string",
                        "example": 1
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}